requirements.txt

# data handling 
pyYAML
numpy 
pandas 
tqdm

# time series processing
tsml_eval
aeon
tsfresh

# classification models
torch
scikit-learn

# plotting and visualization
matplotlib
seaborn
ipywidgets

# testing
pytest
pytest-cov

# mixed effect analysis
statsmodels